Decimal conversion for more accurate level information

+++ Kaguya/Kaguya/Discord/Commands/Exp/Profile.cs
@@ -51,7 +51,7 @@ namespace Kaguya.Discord.Commands.Exp
             
             int fishCount = await _fishRepository.CountAllNonTrashAsync(user.UserId);
 
-            double percent = user.PercentToNextLevel;
+            decimal percent = user.PercentToNextLevel;
 
             IEmote diamondsEmote = _commonEmotes.KaguyaDiamondsAnimated;
